refactor: move shared helpers
