+++ app/src/espresso/kotlin/xyz/lbres/trickcalculator/ui/calculator/calculatorHelpers.kt
@@ -1,10 +1,10 @@
 package xyz.lbres.trickcalculator.ui.calculator
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import xyz.lbres.trickcalculator.R
+import xyz.lbres.trickcalculator.testutils.clearText
 import xyz.lbres.trickcalculator.testutils.closeFragment
 import xyz.lbres.trickcalculator.testutils.openAttributionsFragment
 import xyz.lbres.trickcalculator.testutils.textsaver.TextSaver.Companion.clearSavedText
@@ -13,55 +13,6 @@ import xyz.lbres.trickcalculator.testutils.textsaver.TextSaver.Companion.saveTex
 import xyz.lbres.trickcalculator.testutils.withAnyText
 
 private val mainText = onView(withId(R.id.mainText))
-
-/**
- * Type text in main textview by clicking appropriate button for each character.
- *
- * @param text [String]: text to type
- */
-fun typeText(text: String) {
-    for (c in text) {
-        val buttonId = when (c) {
-            '1' -> R.id.oneButton
-            '2' -> R.id.twoButton
-            '3' -> R.id.threeButton
-            '4' -> R.id.fourButton
-            '5' -> R.id.fiveButton
-            '6' -> R.id.sixButton
-            '7' -> R.id.sevenButton
-            '8' -> R.id.eightButton
-            '9' -> R.id.nineButton
-            '0' -> R.id.zeroButton
-            '+' -> R.id.plusButton
-            '-' -> R.id.minusButton
-            'x' -> R.id.timesButton
-            '/' -> R.id.divideButton
-            '^' -> R.id.expButton
-            '(' -> R.id.lparenButton
-            ')' -> R.id.rparenButton
-            '.' -> R.id.decimalButton
-            else -> null
-        }
-
-        if (buttonId != null) {
-            onView(withId(buttonId)).perform(click())
-        }
-    }
-}
-
-/**
- * Click clear button
- */
-fun clearText() {
-    onView(withId(R.id.clearButton)).perform(click())
-}
-
-/**
- * Click equals button
- */
-fun equals() {
-    onView(withId(R.id.equalsButton)).perform(click())
-}
 
 /**
  * Check that the main textview matches several different values when repeating test
