+++ app/src/espresso/kotlin/xyz/lbres/trickcalculator/ui/calculator/testLastHistoryItem.kt
@@ -8,6 +8,9 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.hamcrest.Matchers.not
 import xyz.lbres.trickcalculator.R
+import xyz.lbres.trickcalculator.testutils.clearText
+import xyz.lbres.trickcalculator.testutils.equals
+import xyz.lbres.trickcalculator.testutils.typeText
 import xyz.lbres.trickcalculator.testutils.withAnyText
 
 private val mainText = onView(withId(R.id.mainText))
