+++ app/src/espresso/kotlin/xyz/lbres/trickcalculator/ui/history/historyGeneralHelpers.kt
@@ -13,13 +13,13 @@ import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
 import xyz.lbres.trickcalculator.R
+import xyz.lbres.trickcalculator.testutils.clearText
 import xyz.lbres.trickcalculator.testutils.closeFragment
+import xyz.lbres.trickcalculator.testutils.equals
 import xyz.lbres.trickcalculator.testutils.openHistoryFragment
 import xyz.lbres.trickcalculator.testutils.openSettingsFragment
+import xyz.lbres.trickcalculator.testutils.typeText
 import xyz.lbres.trickcalculator.testutils.viewassertions.isNotPresented
-import xyz.lbres.trickcalculator.ui.calculator.clearText
-import xyz.lbres.trickcalculator.ui.calculator.equals
-import xyz.lbres.trickcalculator.ui.calculator.typeText
 import java.math.BigDecimal
 import java.math.MathContext
 import java.math.RoundingMode
